\name{hmdbCompMZ}
\alias{hmdbCompMZ}
\docType{data}
\title{
the m/z values of common adduct ions of HMDB metaboites
}
\description{
Monoisotopic mass of common adduct ions.
}
\usage{data("hmdbCompMZ")}
\format{
  A data frame with 5704350 observations on the following 3 variables.
  \describe{
    \item{\code{HMDB_ID}}{HMDB ID.}
    \item{\code{ion.type}}{Adduct ion type.}
    \item{\code{m.z}}{the m/z of the adduct ion.}
  }
}
\source{
https://hmdb.ca/
}
\references{
https://hmdb.ca/
}
\examples{
data(hmdbCompMZ)
}
\keyword{datasets}
