\name{metapone-package}
\alias{metapone-package}
\docType{package}
\title{
\packageTitle{metapone}
}
\description{
\packageDescription{metapone}
The package conducts a permutation test that is based on the hypergeometric test for metabolic pathway analysis. 
}
\details{
The package conducts a weighted hypergeometric test using permutations on metabolomics data. The weights are assigned based on how many metabolites each data feature can match to. The user can tune a parameter to change the penalty for multiple-matched features. 
}
\author{
Tianwei Yu<yutianwei@cuhk.edu.cn>
}
\keyword{ package }
