\name{neg}
\alias{neg}
\docType{data}
\title{
Example negative mode data
}
\description{
Data matrix from negative mode results.
}
\usage{data("neg")}
\format{
  A data frame with 3337 observations on the following 4 variables.
  \describe{
    \item{\code{m.z}}{a numeric vector}
    \item{\code{retention.time}}{a numeric vector}
    \item{\code{p.value}}{a numeric vector}
    \item{\code{statistic}}{a numeric vector}}
}
\examples{
data(neg)
}
\keyword{datasets}
