\name{pos}
\alias{pos}
\docType{data}
\title{
Example positive mode data
}
\description{
Data matrix from positive mode results.}
\usage{data("pos")}
\format{
  A data frame with 2252 observations on the following 4 variables.
  \describe{
    \item{\code{m.z}}{a numeric vector}
    \item{\code{retention.time}}{a numeric vector}
    \item{\code{p.value}}{a numeric vector}
    \item{\code{statistic}}{a numeric vector}
  }
}
\examples{
data(pos)
}
\keyword{datasets}
