\name{metapone-package}
\alias{metapone-package}
\docType{package}
\title{
\packageTitle{metapone}
}
\description{
\packageDescription{metapone}
The package conducts a permutation test that is based on the hypergeometric test for metabolic pathway analysis. 
}
\details{
The package conducts a weighted hypergeometric test on metabolomics data
}
\author{
Tianwei Yu<yutianwei@cuhk.edu.cn>
}
\keyword{ package }
